PCcontrol/requirement.txt
pyserial==3.5
pynput==1.7.6
PyAutoGUI==0.9.53